# pkg/server/web/openapi.yaml
openapi: "3.0.3"
info:
  title: "lumidive API"
  version: "1.0.0"
  description: "TicketDive event adapter, metadata proxy, batch resolution, and iCalendar feed generator."
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:8080"
    description: "Local development server"

paths:
  /healthz:
    get:
      summary: "Health Check"
      description: "Returns health status and version of lumidive service."
      operationId: "getHealth"
      responses:
        "200":
          description: "Healthy"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/events/{eventId}:
    get:
      summary: "Get Event by ID"
      description: "Fetches structured event metadata from TicketDive by event ID or slug."
      operationId: "getEventById"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          description: "TicketDive event ID or URL slug (e.g. plkt1022)"
          schema:
            type: "string"
        - name: "refresh"
          in: "query"
          required: false
          description: "Bypass cache and force refresh from upstream"
          schema:
            type: "boolean"
            default: false
      responses:
        "200":
          description: "Event details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "400":
          description: "Invalid Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Event Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/events/{eventId}/ical:
    get:
      summary: "Get Event iCalendar Feed"
      description: "Generates an iCalendar (.ics) RFC 5545 calendar feed for the specified event."
      operationId: "getEventIcal"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          description: "TicketDive event ID or URL slug"
          schema:
            type: "string"
      responses:
        "200":
          description: "iCalendar feed"
          content:
            text/calendar:
              schema:
                type: "string"
        "404":
          description: "Event Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/artists/{artistId}:
    get:
      summary: "Get Artist Profile and Events"
      description: "Fetches structured artist profile metadata and scheduled upcoming/active events from TicketDive by artist slug or ID."
      operationId: "getArtistById"
      parameters:
        - name: "artistId"
          in: "path"
          required: true
          description: "TicketDive artist ID or URL slug (e.g. yoruami)"
          schema:
            type: "string"
        - name: "refresh"
          in: "query"
          required: false
          description: "Bypass cache and force refresh from upstream"
          schema:
            type: "boolean"
            default: false
      responses:
        "200":
          description: "Artist details and related events"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDetailResponse"
        "400":
          description: "Invalid Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: "Artist Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/artists/{artistId}/ical:
    get:
      summary: "Get Artist iCalendar Feed"
      description: "Generates an aggregated RFC 5545 iCalendar feed for all scheduled events of the specified artist."
      operationId: "getArtistIcal"
      parameters:
        - name: "artistId"
          in: "path"
          required: true
          description: "TicketDive artist ID or URL slug (e.g. yoruami)"
          schema:
            type: "string"
      responses:
        "200":
          description: "iCalendar feed of artist events"
          content:
            text/calendar:
              schema:
                type: "string"
        "404":
          description: "Artist Not Found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/events:
    get:
      summary: "Get Event by URL Query"
      description: "Fetches structured event metadata by specifying the event URL as a query parameter."
      operationId: "getEventByUrl"
      parameters:
        - name: "url"
          in: "query"
          required: true
          description: "TicketDive URL (full or shortened, e.g. https://ticketdive.com/event/plkt1022 or https://t-dv.com/plkt1022)"
          schema:
            type: "string"
        - name: "refresh"
          in: "query"
          required: false
          description: "Bypass cache and force refresh from upstream"
          schema:
            type: "boolean"
            default: false
      responses:
        "200":
          description: "Event details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "400":
          description: "Invalid Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/events/parse:
    post:
      summary: "Parse Event"
      description: "Accepts a TicketDive URL or ID in the request body and returns structured metadata."
      operationId: "parseEvent"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseRequest"
      responses:
        "200":
          description: "Event details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventResponse"
        "400":
          description: "Invalid Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/events/batch:
    post:
      summary: "Batch Parse Events"
      description: "Resolves and parses multiple TicketDive events in a single call."
      operationId: "batchParseEvents"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchParseRequest"
      responses:
        "200":
          description: "Batch parse results"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchEventResponse"
        "400":
          description: "Invalid Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthResponse:
      type: "object"
      required:
        - "status"
        - "service"
        - "version"
        - "timestamp"
      properties:
        status:
          type: "string"
          example: "ok"
        service:
          type: "string"
          example: "lumidive"
        version:
          type: "string"
          example: "1.0.0"
        timestamp:
          type: "integer"
          format: "int64"
          example: 1791377890202

    ParseRequest:
      type: "object"
      properties:
        url:
          type: "string"
          description: "TicketDive event URL"
          example: "https://ticketdive.com/event/plkt1022"
        id:
          type: "string"
          description: "TicketDive event ID"
          example: "plkt1022"
        refresh:
          type: "boolean"
          description: "Bypass cache and force refresh"
          default: false

    BatchParseRequest:
      type: "object"
      required:
        - "targets"
      properties:
        targets:
          type: "array"
          items:
            type: "string"
          description: "List of event URLs or IDs to parse"
          example: ["plkt1022", "https://t-dv.com/tgg_1006"]
        refresh:
          type: "boolean"
          description: "Bypass cache and force refresh"
          default: false

    SourceMetadata:
      type: "object"
      description: "Upstream source provenance and platform details"
      properties:
        platform:
          type: "string"
          description: "Source ticketing platform"
          example: "ticketdive"
        url:
          type: "string"
          description: "Original source URL"
          example: "https://ticketdive.com/event/plkt1022"
        buildId:
          type: "string"
          description: "Next.js upstream build ID"
          example: "DKPoTbxAqISse_0XBEukK"
        fetchedAt:
          type: "string"
          format: "date-time"
          description: "Upstream fetch timestamp"
        serverTime:
          type: "integer"
          format: "int64"
          description: "Server timestamp reported by upstream"

    EventResponse:
      type: "object"
      required:
        - "apiVersion"
        - "success"
        - "cached"
        - "data"
      properties:
        apiVersion:
          type: "string"
          example: "1.0"
        success:
          type: "boolean"
          example: true
        cached:
          type: "boolean"
          example: false
        cachedAt:
          type: "string"
          format: "date-time"
          description: "Timestamp when the cache entry was created"
        source:
          $ref: "#/components/schemas/SourceMetadata"
        data:
          $ref: "#/components/schemas/Event"

    ArtistDetailResponse:
      type: "object"
      required:
        - "apiVersion"
        - "success"
        - "cached"
        - "data"
      properties:
        apiVersion:
          type: "string"
          example: "1.0"
        success:
          type: "boolean"
          example: true
        cached:
          type: "boolean"
          example: false
        cachedAt:
          type: "string"
          format: "date-time"
        source:
          $ref: "#/components/schemas/SourceMetadata"
        data:
          $ref: "#/components/schemas/ArtistDetail"

    ArtistDetail:
      type: "object"
      required:
        - "id"
        - "name"
      properties:
        id:
          type: "string"
          description: "Artist slug or ID"
          example: "yoruami"
        name:
          type: "string"
          description: "Artist display name"
          example: "夜光性アミューズ"
        profileImage:
          type: "string"
          description: "Artist profile avatar image URL"
        twitterAccount:
          type: "string"
          description: "Official Twitter / X account handle"
          example: "Yoruamiofficial"
        url:
          type: "string"
          description: "Official TicketDive artist page URL"
          example: "https://ticketdive.com/artist/yoruami"
        events:
          type: "array"
          items:
            $ref: "#/components/schemas/ArtistEvent"
        metadata:
          type: "object"
          additionalProperties: true

    ArtistEvent:
      type: "object"
      required:
        - "id"
        - "title"
      properties:
        id:
          type: "string"
          description: "Event ID"
        slug:
          type: "string"
          description: "Event slug URL identifier"
        title:
          type: "string"
          description: "Event title"
        url:
          type: "string"
          description: "Event URL on TicketDive"
        imageSource:
          type: "string"
          description: "Event flyer/banner thumbnail URL"
        startAt:
          type: "string"
          description: "Event start or stage date (ISO 8601)"
        endAt:
          type: "string"
          description: "Event end date if specified"
        venueName:
          type: "string"
          description: "Venue name"
        salesStatus:
          type: "string"
          description: "Current sales status (applied, closed, etc.)"

    BatchEventResult:
      type: "object"
      required:
        - "target"
        - "success"
      properties:
        target:
          type: "string"
          description: "Requested target URL or ID"
        success:
          type: "boolean"
        event:
          $ref: "#/components/schemas/Event"
        error:
          type: "string"
          description: "Error message if parsing this target failed"

    BatchEventResponse:
      type: "object"
      required:
        - "apiVersion"
        - "success"
        - "results"
      properties:
        apiVersion:
          type: "string"
          example: "1.0"
        success:
          type: "boolean"
          example: true
        results:
          type: "array"
          items:
            $ref: "#/components/schemas/BatchEventResult"

    Event:
      type: "object"
      required:
        - "id"
        - "slug"
        - "name"
        - "url"
        - "shortUrl"
      properties:
        id:
          type: "string"
          description: "Internal event ID (e.g. Firestore document ID)"
          example: "drIPQl23ZVboh1BWQy6y"
        slug:
          type: "string"
          description: "Event URL slug identifier"
          example: "plkt1022"
        name:
          type: "string"
          description: "Event title"
          example: "2026/10/22(木) 「笑う門にはプリュきたる!! 」@WOMBLIVE"
        detail:
          type: "string"
          description: "Event description and notes"
        release:
          type: "string"
          description: "Information release date"
          example: "2026-09-28T13:00:00.000Z"
        url:
          type: "string"
          description: "Canonical event URL"
          example: "https://ticketdive.com/event/plkt1022"
        shortUrl:
          type: "string"
          description: "Shortened event URL"
          example: "https://t-dv.com/plkt1022"
        stages:
          type: "array"
          items:
            $ref: "#/components/schemas/Stage"
        artists:
          type: "array"
          items:
            $ref: "#/components/schemas/Artist"
        ticketGroups:
          type: "array"
          items:
            $ref: "#/components/schemas/TicketGroup"
        images:
          type: "array"
          items:
            type: "string"
          description: "Flyer and banner image URLs"
        stats:
          $ref: "#/components/schemas/EventStats"
        metadata:
          type: "object"
          additionalProperties: true
          description: "Extensible dictionary for future or platform-specific attributes"

    EventStats:
      type: "object"
      description: "Aggregated statistical indicators for the event"
      properties:
        totalTicketTypes:
          type: "integer"
          description: "Total number of distinct ticket types"
          example: 10
        soldOutTicketTypes:
          type: "integer"
          description: "Number of sold out ticket types"
          example: 5
        minPrice:
          type: "integer"
          description: "Lowest available ticket price"
          example: 1000
        maxPrice:
          type: "integer"
          description: "Highest available ticket price"
          example: 4200
        hasAvailableTickets:
          type: "boolean"
          description: "Whether at least one ticket type remains available"
          example: true

    Stage:
      type: "object"
      required:
        - "id"
        - "stageName"
      properties:
        id:
          type: "string"
          example: "RyL58qx8oMmAtzapdEZY"
        stageName:
          type: "string"
          example: "WOMBLIVE"
        startAt:
          type: "string"
          description: "Show start time (ISO 8601)"
          example: "2026-10-22T06:45:00.000Z"
        openAt:
          type: "string"
          description: "Doors open time (ISO 8601)"
          example: "2026-10-22T06:15:00.000Z"
        endAt:
          type: "string"
          description: "Estimated or scheduled finish time (ISO 8601)"
          example: "2026-10-22T10:00:00.000Z"
        venue:
          $ref: "#/components/schemas/Venue"
        artists:
          type: "array"
          items:
            $ref: "#/components/schemas/Artist"
        timetable:
          type: "array"
          items:
            $ref: "#/components/schemas/TimetableSlot"
          description: "Optional slot-by-slot schedule if resolved"
        metadata:
          type: "object"
          additionalProperties: true
          description: "Extensible stage metadata"

    Venue:
      type: "object"
      properties:
        name:
          type: "string"
          example: "WOMBLIVE"
        address:
          type: "string"
          example: "東京都渋谷区円山町2-16"
        url:
          type: "string"
          description: "Venue official website or map URL"
        metadata:
          type: "object"
          additionalProperties: true

    Artist:
      type: "object"
      required:
        - "id"
        - "name"
      properties:
        id:
          type: "string"
          description: "Artist slug or ID"
          example: "honeyspice-re"
        name:
          type: "string"
          description: "Artist display name"
          example: "ハニースパイスRe."
        url:
          type: "string"
          description: "Profile or TicketDive artist URL"
          example: "https://ticketdive.com/artist/honeyspice-re"
        metadata:
          type: "object"
          additionalProperties: true

    TimetableSlot:
      type: "object"
      required:
        - "artistName"
      properties:
        artistId:
          type: "string"
        artistName:
          type: "string"
        startAt:
          type: "string"
          format: "date-time"
        endAt:
          type: "string"
          format: "date-time"
        parallelOrder:
          type: "integer"
          description: "Stage or parallel group index"
        metadata:
          type: "object"
          additionalProperties: true

    TicketGroup:
      type: "object"
      required:
        - "id"
        - "name"
        - "receptionType"
        - "types"
      properties:
        id:
          type: "string"
          example: "EHV68qpqSYQB2qdJGAji"
        name:
          type: "string"
          description: "Group name (e.g. 一般販売, 最前OP 一般販売, FC先行販売)"
          example: "一般販売"
        receptionType:
          type: "string"
          description: "Reception method (first=先着, lottery=抽選)"
          example: "first"
        startApply:
          type: "string"
          description: "Application start timestamp (ISO 8601)"
          example: "2026-10-02T11:00:00.000Z"
        endApply:
          type: "string"
          description: "Application end timestamp (ISO 8601)"
          example: "2026-10-22T06:45:00.000Z"
        paymentChannels:
          type: "array"
          items:
            type: "string"
          example: ["card", "konbini"]
        restrictions:
          type: "array"
          items:
            type: "string"
          description: "Rules such as untransferable, fcOnly, phoneNumberNeeded"
          example: ["untransferable"]
        customize:
          type: "array"
          items:
            $ref: "#/components/schemas/CustomizeQuestion"
        status:
          type: "string"
          description: "Current reception status (e.g. applied, closed, upcoming)"
          example: "applied"
        types:
          type: "array"
          items:
            $ref: "#/components/schemas/TicketType"
        metadata:
          type: "object"
          additionalProperties: true

    CustomizeQuestion:
      type: "object"
      properties:
        id:
          type: "string"
        label:
          type: "string"
          example: "お目当ての出演者を教えてください（必須）"
        required:
          type: "boolean"
          example: true
        type:
          type: "string"
          example: "select"
        options:
          type: "array"
          items:
            type: "string"
        metadata:
          type: "object"
          additionalProperties: true

    TicketType:
      type: "object"
      required:
        - "id"
        - "name"
        - "price"
        - "isSoldOut"
      properties:
        id:
          type: "string"
          example: "5ekMbDP1Vhc1qgIt6etj"
        name:
          type: "string"
          example: "一般前方チケット"
        detail:
          type: "string"
          example: "ドリンク代別700円込み/6歳以上チケット必要"
        currency:
          type: "string"
          description: "Three-letter ISO 4217 currency code"
          default: "JPY"
          example: "JPY"
        price:
          type: "integer"
          description: "Base ticket price"
          example: 4200
        fee:
          type: "integer"
          description: "Handling fee"
          example: 340
        totalPrice:
          type: "integer"
          description: "Combined price and fee"
          example: 4540
        remainingRate:
          type: "number"
          description: "Stock ratio (1.0 = in stock, 0.0 = sold out)"
          example: 1.0
        isSoldOut:
          type: "boolean"
          description: "Whether this ticket type is currently sold out"
          example: false
        status:
          type: "string"
          description: "Status string from upstream (applied, closed, etc.)"
          example: "applied"
        maxNumPerApply:
          type: "integer"
          description: "Maximum tickets allowed per application"
          example: 4
        prefix:
          type: "string"
          description: "Ticket reference number prefix (e.g. IA, IB, IHS)"
          example: "IA"
        metadata:
          type: "object"
          additionalProperties: true
          description: "Extensible vendor-specific ticket attributes"

    ErrorResponse:
      type: "object"
      required:
        - "apiVersion"
        - "success"
        - "error"
      properties:
        apiVersion:
          type: "string"
          example: "1.0"
        success:
          type: "boolean"
          example: false
        code:
          type: "string"
          description: "Machine-readable error classification code"
          example: "EVENT_NOT_FOUND"
        error:
          type: "string"
          description: "Human-readable error description"
          example: "Event not found"
        details:
          type: "array"
          items:
            type: "string"
          description: "Additional diagnostic information or validation errors"